% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemtcTest.r
\name{gemtcTest}
\alias{gemtcTest}
\title{A Cat Function}
\usage{
gemtcTest()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
gemtcTest()
}
\keyword{cats}
